requirements.txt
flask
pyautogui
openai
python-dotenv
dspy
boto3
moondream
Pillow
